fix tests; org files; proofs optional; do not need both inference + forecast
