+++ x/emissions/module/rewards/rewards_test.go
@@ -1068,6 +1068,7 @@ func (s *RewardsTestSuite) TestStandardRewardEmissionWithOneInfererAndOneReputer
 	worker1Sig, err := GenerateWorkerSignature(s, worker1InferenceForecastBundle, worker)
 	s.Require().NoError(err)
 	worker1Bundle := &types.WorkerDataBundle{
+		Worker:                             worker.String(),
 		InferenceForecastsBundle:           worker1InferenceForecastBundle,
 		InferencesForecastsBundleSignature: worker1Sig,
 		Pubkey:                             GetAccPubKey(s, worker),
